ProjetoPI-final/requirements.txt
Flask==3.0.3
python-dotenv==1.0.0
openai==1.14.3
mysql-connector-python==8.3.0
requests==2.31.0
pytest==8.2.1
pytest-mock==3.14.0
pytest-html==4.1.1
pytest-cov==6.1.1




